+++ src/checker/providers/detectProvider.ts
@@ -1,7 +1,9 @@
-import { ConfidenceLevel } from "../../enumerations";
+import { ConfidenceLevel, StatsEventType } from "../../enumerations";
 
 export interface IDetectProvider {
     readonly name: string;
+    readonly eventWhenTechnical: StatsEventType;
+    readonly eventWhenMessage: StatsEventType;
     readonly isStopOnEval: boolean;
     
     check(text: string): [ConfidenceLevel, string];