+++ src/checker/providers/entropyDetect.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { IDetectProvider } from "./detectProvider";
-import { ConfidenceLevel } from "../../enumerations";
-import { Constants } from '../../constants';
+import { ConfidenceLevel, StatsEventType } from "../../enumerations";
+import { Constants, ProviderName } from '../../constants';
 import { stripNonAlphaCharacters } from '../../helpers/utils';
 
 export class EntropyDetect implements IDetectProvider {
@@ -11,8 +11,11 @@ export class EntropyDetect implements IDetectProvider {
         this._threshold = vscode.workspace.getConfiguration(Constants.ExtensionID).get<number>('entropy.threshold')!;
     }
 
-    readonly name: string = 'Entropy provider';
+    //#region IDetectProvider
 
+    readonly name: string = ProviderName.Entropy;
+    readonly eventWhenTechnical: StatsEventType = StatsEventType.DetectedAsTechnicalByEntropy;
+    readonly eventWhenMessage: StatsEventType = StatsEventType.DetectedAsMessageByEntropy;
     readonly isStopOnEval: boolean = false;
 
     check(text: string): [ConfidenceLevel, string] {
@@ -20,6 +23,8 @@ export class EntropyDetect implements IDetectProvider {
 
         return [entropy > this._threshold ? ConfidenceLevel.Message : ConfidenceLevel.Unknown, `entropy=${entropy.toFixed(2)}`];
     }
+
+    //#endregion
 
     private getShannonEntropy(text: string): number {
         let map: {[id: string]: number; } = {};