+++ src/checker/providers/stringDetect.ts
@@ -1,12 +1,18 @@
 import { IDetectProvider } from "./detectProvider";
-import { ConfidenceLevel } from "../../enumerations";
+import { ConfidenceLevel, StatsEventType } from "../../enumerations";
+import { ProviderName } from "../../constants";
 
 export class StringDetect implements IDetectProvider {
-    readonly name: string = 'String provider';
+    //#region IDetectProvider
 
+    readonly name: string = ProviderName.String;
+    readonly eventWhenTechnical: StatsEventType = StatsEventType.DetectedAsTechnicalByString;
+    readonly eventWhenMessage: StatsEventType = StatsEventType.DetectedAsMessageByString;
     readonly isStopOnEval: boolean = false;
 
     check(_text: string): [ConfidenceLevel, string] {
         return [ConfidenceLevel.Message, '*'];
     }
+
+    //#endregion
 }