+++ src/stats/statsEmiter.ts
@@ -1,0 +1,5 @@
+import { StatsEventType } from "../enumerations";
+
+export interface IStatsEmiter {    
+    emit(event: StatsEventType, value?: number): void;
+}